asdf

+++ hw1/sieve_of_erastosthenes_b.cpp
@@ -5,7 +5,7 @@
 #define n 3
 
 //function finds remainder with the k,n,B, and divisor inputs (replaces remainder with right value)
-double findingremainder(double k[n],double d[n], double B, double divisor,double &remainder)
+double findingremainder(double k[n],double d[n], double B, int divisor,double &remainder)
 {
 	for (int i=0;i<n;i++)
 
@@ -24,7 +24,7 @@ main()
 	double k[n] = {0, 1, 2};
 	double d[n] = {7, 5, 2};
 	double B = 10;
-	double divisor = 3;
+	int divisor = 3;
 	double remainder = 0;
 	
 	findingremainder(k,d, B,divisor,remainder);
